moved the functions out of the modal.
