+++ nutrinova-ui/src/app/(authorized)/patients/page.tsx
@@ -1,14 +1,18 @@
 'use client'
-import { Box, IconButton, List, ListItem, Typography } from "@mui/material";
+import { Box, Button, IconButton, List, ListItem, Typography } from "@mui/material";
 import { PatientInfoModal } from "@/app/(authorized)/patients/_components/PatientInfoModal";
 import { PageContainer } from "@/components/PageContainer";
 import { useContext, useState } from "react";
 import { PatientContext } from "@/components/providers/PatientProvider";
 import { Delete } from "@mui/icons-material";
-import { Patient } from "./_models/patient";
+import { CreatePatientReq, Patient } from "./_models/patient";
+import { PatientForm } from "./_components/PatientInfoForm";
+import { useCreatePatientMutation } from '@/app/(authorized)/patients/patientHooks';
 
 const PatientsPage = () => {
   const { patients } = useContext(PatientContext);
+  const createPatientMutation = useCreatePatientMutation();
+
 
   const handleDelete = (patient: Patient) => {
     // TODO: how should we handle this?
@@ -22,12 +26,25 @@ const PatientsPage = () => {
     setOpen(!open);
   }
 
+  const HandlePatientAdd = (patientInfo: PatientForm) => {
+    const patient: CreatePatientReq = {
+      firstname: patientInfo.name.split(' ')[0],
+      lastname: patientInfo.name.split(' ')[1] ?? '',
+      sex: patientInfo?.sex,
+      base64image: patientInfo?.pff,
+      age: patientInfo?.age,
+      useDefaultNutrientGoals: patientInfo.optOut
+    }
+    createPatientMutation.mutate(patient);
+  }
+
+
   return (
     <>
       <PageContainer title="Patients">
         <Box>
-          <Typography variant="button">Add a Patient</Typography>
-          <PatientInfoModal openModal={open} onClose={toggleOpen} />
+          <Button onClick={toggleOpen}>Add a Patient</Button>
+          <PatientInfoModal openModal={open} onClose={toggleOpen} submitFunction={HandlePatientAdd} />
         </Box >
 
         <Typography variant="button">Patients</Typography>
